requirements.txt
alembic==1.16.2
anyio==4.9.0
bcrypt==4.3.0
click==8.2.1
databases==0.9.0
email_validator==2.2.0
fastapi==0.115.13
greenlet==3.2.3
httptools==0.6.4
idna==3.10
mysql-connector-python==9.3.0
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
SQLAlchemy==2.0.41
uvicorn==0.34.3
watchfiles==1.1.0
websockets==10.4
